Add migration for tokens table to harden "legacy" API tokens
Legacy API tokens are tokens which never expire, are persisted (as
opposed to ephemeral tokens which confirm to the access/refresh model),
and can only be revoked by removing it from the db.

This commit hardens those legacy tokens by encrypting the token at rest
and using its sha256 hash to determine if it should be honored.