Attach errors to org user association ops; allow arbitrary user <> org association by user_id
We can consider if requiring the invitation token mechanism is the only
viable way a user can be added to an org... this would require the user
to have an emai already registered or they would receive the token
out-of-band and then confirm the invite. This commit relaxes the api to
allow any arbitrary user_id to be associated with the org. This is a
good chance to document this behavior once and for all.